ldap_mysql_granter/prod_requirements.txt
boto==2.38.0
Jinja2==2.7.3
MarkupSafe==0.23
mccabe==0.3
MySQL-python==1.2.5
python-ldap==2.4.19
PyYAML==3.11
requests==2.6.0
